requirements.txt
# --- FastAPI Core ---
fastapi==0.110.2
uvicorn[standard]==0.29.0

# --- File Uploads & CORS ---
python-multipart==0.0.9
aiofiles==23.2.1
requests==2.32.2
httpx==0.27.0

# --- Auth & Security ---
python-jose[cryptography]==3.3.0

# --- Task Queue ---
celery==5.3.6
redis==5.0.3

# --- Cloud + Twilio ---
twilio==9.0.4
cloudinary==1.38.0

# --- AI Insight & Env ---
openai==1.30.3
python-dotenv==1.0.1

# --- Inference ---
ultralytics==8.1.25                  
Pillow==10.3.0
certifi==2024.6.2
numpy==1.26.4
